// src/types/user.ts
export type UsersCustom = {
  id: string;
  email: string | null;
  phone: string | null;
  full_name: string | null;
  q1: string | null;
  q2: string | null;
  q3: string | null;
  image1_path: string | null;
  image2_path: string | null;
  image3_path: string | null;
  image4_path: string | null;
  image5_path: string | null;
  image6_path: string | null;
  image7_path: string | null;
  created_at?: string | null;
};